Experimental mocha/selenium testing.
In my research I have found no simple way of using Jest with the current browser
test running tech. It's either Jasmine or Mocha that people seem to be using.

Initially I wanted to go with the Karma test runner using karma-firefox-launcher
but I could not get it to work with ff v61 and higher and I'm unwilling to
invest time in that until they get that fixed.

This commit uses the webextensions-geckodriver package to leverage selenium and
webdriver. Currently I haven't had much success in interacting with tridactyl in
this form but it does work.